added: "new" state and confirm to delete

+++ packages/server/public/src/components/command-bar/commands/handle-file-delete.ts
@@ -5,6 +5,19 @@ export const handleFileDelete: CommandHandler = async ({ context }) => ({
     const { id } = context.routeService.getNoteConfigFromUrl();
     if (id) {
       try {
+        const data = await context.apiService.getIncomingLinks(id);
+        if (data.incomingLinks.length) {
+          const isConfirmed = context.windowRef.window.confirm(
+            `This note is referenced by other notes. Are you sure you want to delete it?\n\nNotes that are referencing this note:\n` +
+              data.incomingLinks.map((item) => item.title).join("\n")
+          );
+
+          if (!isConfirmed) {
+            context.notificationService.displayMessage("Delete canceled", "info");
+            return;
+          }
+        }
+
         await context.apiService.deleteNote(id);
         await context.syncService.syncAllFileVersions();
         context.notificationService.displayMessage("Note deleted", "warning");
